make file

+++ client_src/exwraper.c
@@ -1,11 +1,18 @@
-//
-// Created by Robert JONES on 2017/07/17.
-//
-
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exwraper.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rojones <marvin@42.fr>                     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2017/07/26 07:34:57 by rojones           #+#    #+#             */
+/*   Updated: 2017/07/26 07:40:25 by rojones          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "irc_client.h"
 
-size_t x_size_t(int err, size_t res, char *str, char *file, int line)
+size_t	x_size_t(size_t err, size_t res, char *str, char *file, int line)
 {
 	if (res == err)
 	{
@@ -17,7 +24,7 @@ size_t x_size_t(int err, size_t res, char *str, char *file, int line)
 	return (res);
 }
 
-int x_int(int err, int res, char *str, char *file, int line)
+int		x_int(int err, int res, char *str, char *file, int line)
 {
 	if (res == err)
 	{
@@ -41,7 +48,7 @@ long	x_long(int err, long res, char *str, char *file, int line)
 	return (res);
 }
 
-void *x_void(void *err, void *res, char *str, char *file, int line)
+void	*x_void(void *err, void *res, char *str, char *file, int line)
 {
 	if (res == err)
 	{
@@ -53,7 +60,7 @@ void *x_void(void *err, void *res, char *str, char *file, int line)
 	return (res);
 }
 
-void print_err_exit(char *class, char *file, int line, char *msg)
+void	print_err_exit(char *class, char *file, int line, char *msg)
 {
 	close_all();
 	fprintf(stderr, "%s error (%s, %d): %s\n",
